[rename] match -> match/like 분리
