+++ src/main/java/com/example/demo/src/match/MatchService.java
@@ -1,9 +1,6 @@
 package com.example.demo.src.match;
 
 import com.example.demo.config.BaseException;
-import com.example.demo.src.match.model.PostTeamLikeRes;
-import com.example.demo.src.match.model.PostUserLikeRes;
-import com.example.demo.src.match.model.PostUserPassRes;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -21,125 +18,6 @@ public class MatchService {
 
     private final MatchProvider matchProvider;
     private final MatchDao matchDao;
-
-    public void updateUserLike(int userLikeIdx) throws BaseException {
-        try {
-            int result = matchDao.updateUserLike(userLikeIdx);
-            if(result == 0){
-                throw new BaseException(FAIL_USERLIKE);
-            }
-        } catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public PostUserLikeRes createUserLike(int likerIdx, int likingIdx) throws BaseException {
-        // likingIdx 유효성 검사
-        if(matchProvider.checkUserIdx(likingIdx) == 0){
-            throw new BaseException(INVALID_USERIDX);
-        }
-
-        // 중복 좋아요 유효성 검사
-        if(matchProvider.checkPreUserLike(likerIdx, likingIdx) == 1){
-            throw new BaseException(USERLIKE_INVALID_LIKINGIDX);
-        }
-
-        try {
-            // 이전에 비활성화나 좋아요 취소를 했던 유저의 경우
-            int preUserLikeIdx = matchProvider.checkUserLikeHist(likerIdx, likingIdx);
-            if(preUserLikeIdx != 0){
-                // active 업데이트
-                updateUserLike(preUserLikeIdx);
-                return new PostUserLikeRes(preUserLikeIdx);
-            }
-
-            int userLikeIdx = matchDao.createUserLike(likerIdx, likingIdx);
-            return new PostUserLikeRes(userLikeIdx);
-        } catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public void deleteUserLike(int likerIdx, int likingIdx) throws BaseException {
-        // likingIdx 유효성 검사
-        if(matchProvider.checkUserIdx(likingIdx) == 0){
-            throw new BaseException(INVALID_USERIDX);
-        }
-
-        // 좋아요 누르지 않은 유저에 대해 좋아요 취소 요청을 할 경우
-        if(matchProvider.checkPreUserLike(likerIdx, likingIdx) == 0){
-            throw new BaseException(UNUSERLIKE_INVALID_LIKINGIDX);
-        }
-
-        try {
-            int result = matchDao.deleteUserLike(likerIdx, likingIdx);
-            if(result == 0){
-                throw new BaseException(FAIL_UNUSERLIKE);
-            }
-        } catch (Exception exception){
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public void updateUserPass(int userPassIdx) throws BaseException {
-        try {
-            int result = matchDao.updateUserPass(userPassIdx);
-            if(result == 0){
-                throw new BaseException(FAIL_USERPASS);
-            }
-        } catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public PostUserPassRes createUserPass(int passerIdx, int passingIdx) throws BaseException {
-        // passingIdx 유효성 검사
-        if(matchProvider.checkUserIdx(passingIdx) == 0){
-            throw new BaseException(INVALID_USERIDX);
-        }
-
-        // 중복 넘기기 유효성 검사
-        if(matchProvider.checkPreUserPass(passerIdx, passingIdx) == 1){
-            throw new BaseException(USERPASS_INVALID_PASSINGIDX);
-        }
-
-        try {
-            // 이전에 비활성화나 넘기기 취소를 했던 유저의 경우
-            int preUserPassIdx = matchProvider.checkUserPassHist(passerIdx, passingIdx);
-            if(preUserPassIdx != 0){
-                // active 업데이트
-                updateUserPass(preUserPassIdx);
-                return new PostUserPassRes(preUserPassIdx);
-            }
-
-            int userPassIdx = matchDao.createUserPass(passerIdx, passingIdx);
-            return new PostUserPassRes(userPassIdx);
-        } catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public void deleteUserPass(int passerIdx, int passingIdx) throws BaseException {
-        // passingIdx 유효성 검사
-        if(matchProvider.checkUserIdx(passingIdx) == 0){
-            throw new BaseException(INVALID_USERIDX);
-        }
-
-        // 넘기기를 누르지 않은 유저에 대해 넘기기 취소 요청을 할 경우
-        if(matchProvider.checkPreUserPass(passerIdx, passingIdx) == 0){
-            throw new BaseException(UNUSERPASS_INVALID_PASSINGIDX);
-        }
-
-        try {
-            int result = matchDao.deleteUserPass(passerIdx, passingIdx);
-            if(result == 0){
-                throw new BaseException(FAIL_UNUSERPASS);
-            }
-        } catch (Exception exception){
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
 
     public void createAgeRangeFilter(int userIdx, List<String> ageRange) throws BaseException {
         try {
@@ -350,44 +228,6 @@ public class MatchService {
                 }
             }
         } catch (Exception exception){
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public void updateTeamLike(int teamLikeIdx) throws BaseException {
-        try {
-            int result = matchDao.updateTeamLike(teamLikeIdx);
-            if(result == 0){
-                throw new BaseException(FAIL_TEAMLIKE);
-            }
-        } catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    public PostTeamLikeRes createTeamLike(int likerIdx, int likingIdx) throws BaseException {
-        // likingIdx 유효성 검사
-        if(matchProvider.checkTeamIdx(likingIdx) == 0){
-            throw new BaseException(INVALID_TEAMIDX);
-        }
-
-        // 중복 좋아요 유효성 검사
-        if(matchProvider.checkPreTeamLike(likerIdx, likingIdx) == 1){
-            throw new BaseException(TEAMLIKE_INVALID_LIKINGIDX);
-        }
-
-        try {
-            // 이전에 비활성화나 좋아요 취소를 했던 팀의 경우
-            int preTeamLikeIdx = matchProvider.checkTeamLikeHist(likerIdx, likingIdx);
-            if(preTeamLikeIdx != 0){
-                // active 업데이트
-                updateTeamLike(preTeamLikeIdx);
-                return new PostTeamLikeRes(preTeamLikeIdx);
-            }
-
-            int teamLikeIdx = matchDao.createTeamLike(likerIdx, likingIdx);
-            return new PostTeamLikeRes(teamLikeIdx);
-        } catch (Exception exception) {
             throw new BaseException(DATABASE_ERROR);
         }
     }
